perf: use simd sha256 for cacheContext checksumming
