+++ exporter/local/fs.go
@@ -15,6 +15,7 @@ import (
 	"github.com/docker/docker/pkg/idtools"
 	intoto "github.com/in-toto/in-toto-golang/in_toto"
 	"github.com/moby/buildkit/cache"
+	"github.com/moby/buildkit/depot"
 	"github.com/moby/buildkit/exporter"
 	"github.com/moby/buildkit/exporter/attestation"
 	"github.com/moby/buildkit/exporter/util/epoch"
@@ -22,7 +23,6 @@ import (
 	"github.com/moby/buildkit/snapshot"
 	"github.com/moby/buildkit/solver/result"
 	"github.com/moby/buildkit/util/staticfs"
-	digest "github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
 	"github.com/tonistiigi/fsutil"
 	fstypes "github.com/tonistiigi/fsutil/types"
@@ -149,7 +149,7 @@ func CreateFS(ctx context.Context, sessionID string, k string, ref cache.Immutab
 				return err
 			}
 			defer f.Close()
-			d := digest.Canonical.Digester()
+			d := depot.NewFastDigester()
 			if _, err := io.Copy(d.Hash(), f); err != nil {
 				return err
 			}
